split test/remove println
