+++ aptos-move/e2e-move-tests/src/tests/mod.rs
@@ -9,6 +9,7 @@ mod code_publishing;
 mod common;
 mod constructor_args;
 mod error_map;
+mod fee_payer;
 mod fungible_asset;
 mod gas;
 mod generate_upgrade_script;
